+++ .gitignore
@@ -15,6 +15,8 @@ __pycache__
 .env.local
 .env.*.local
 
+
+
 # Log files
 npm-debug.log*
 yarn-debug.log*
