+++ wsgi.py
@@ -1,4 +1,4 @@
-from visMOP import app
+from visMOP.main import app
 
 if __name__ == "__main__":
     app.run(debug=True)
